\name{FFieldPtRepDemo}
\alias{FFieldPtRepDemo}
\title{Label placement using force field simulation}
\description{
Demonstrates force field simulation 
of interaction of set of points to place labels on a scatterplot.
}
\usage{
FFieldPtRepDemo()
}
\details{
Points experience repulsion from one another and attraction to
their original positions.
Coordinates are normalized and unadjusted and adjusted plots provided.
}
\examples{
FFieldPtRepDemo()          
}
\seealso{
\link{FField-package}
\code{\link{FFieldPtRep}}
}
